requirements.txt
Django>=3.2,<3.3
gunicorn>=20.1.0,<20.2
djangorestframework
attrs
django-filter
django-cors-headers
djongo
dnspython
certifi